Update envoy_reader.py
This works with 0.94.4 homeassistant and older Envoy C,  with R3.7 s/w  (no .json data).   I'm not sure how to use the async requests working, so the import and await statements need fixing.   Otherwise this should still work with newer Envoys, but please test it.   

the envoy_reader.py file is changed obviously,  and the README.md file is also updated.       

Thanks!